Códigos/requirements.txt
Flask==3.1.0
PyMySQL==1.1.2
python-dotenv==1.0.1
